// sdk/ng-sdk-js/examples/multi-framework-signals/src/shapes/ldo/catShape.typings.ts
export type IRI = string;

/**
 * =============================================================================
 * Typescript Typings for catShape
 * =============================================================================
 */

/**
 * Cat Type
 */
export interface Cat {
  id: IRI;
  /**
   * Original IRI: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
   */
  type: "Cat";
  /**
   * Original IRI: http://example.org/name
   */
  name: string;
  /**
   * Original IRI: http://example.org/age
   */
  age: number;
  /**
   * Original IRI: http://example.org/numberOfHomes
   */
  numberOfHomes: number;
  /**
   * Original IRI: http://example.org/address
   */
  address: {
    id: IRI;
    /**
     * Original IRI: http://example.org/street
     */
    street: string;
    /**
     * Original IRI: http://example.org/houseNumber
     */
    houseNumber: string;
    /**
     * Original IRI: http://example.org/floor
     */
    floor: number;
  };
}